Progressive dificult chunk system

+++ Assets/Scripts/TilesManager.cs
@@ -5,15 +5,21 @@ using UnityEngine;
 
 public class TilesManager : MonoBehaviour
 {
-	public GameObject[] tilePrefabs;
+	[SerializeField]
+	private GameObject _initialPrefabs;
+	
 	private List<GameObject> _savedTiles;
 
 	private Transform	_playerTransform;
 	private float		_spawnZ = 0.0f;
 	private float		_tileLength = 20.5f;
-	private float		_safeZone = 18f;
+	private float		_safeZone = 58f;
 	private int			_amountTiles = 8;
 	private int			_lastPrefabIndex = 0;
+
+	private float		_countToDelete = 0;
+
+	public static Dificulties currentDificultChunk = Dificulties.EASY;
 
 	// Use this for initialization
 	void Start()
@@ -35,35 +41,53 @@ public class TilesManager : MonoBehaviour
 	// Update is called once per frame
 	void Update()
 	{
-		if (_playerTransform.position.z - _safeZone > (_spawnZ - _amountTiles * _tileLength))
+		if (_countToDelete < 6)
 		{
-			SpawnTiles();
-			DeleteTiles();
+			Debug.Log(_countToDelete);
+			_countToDelete += Time.deltaTime;
+		}
+		else
+		{
+			if (_playerTransform.position.z - _safeZone > (_spawnZ - _amountTiles * _tileLength))
+			{
+				SpawnTiles();
+				DeleteTiles();
+			}
 		}
 	}
 
+	/// <summary>
+	/// Spawn the first and the second Chunk;
+	/// </summary>
 	private void SpawnInitialTiles()
 	{
-		//throw new NotImplementedException();
+		GameObject go;
+
+		go = Instantiate(_initialPrefabs);
+		go.transform.SetParent(transform);
+		go.transform.position = Vector3.forward * _spawnZ;
+		_spawnZ += _tileLength;
 	}
 
+	/// <summary>
+	/// Spawn a Chunk from the Pool System
+	/// </summary>
 	private void SpawnTiles()
 	{
 		GameObject go;
 
-		go = PoolSystem.GetChunkFromPool(Dificulties.EASY);
+		go = PoolSystem.GetChunkFromPool(currentDificultChunk);
 
 		go.transform.SetParent(transform);
 		go.transform.position = Vector3.forward * _spawnZ;
 		_spawnZ += _tileLength;
 		_savedTiles.Add(go);
 	}
-
+	
 	private void DeleteTiles()
 	{
 		PoolSystem.AddChunkToPool(_savedTiles[0]);
+		Debug.Log(_savedTiles[0]);
 		_savedTiles.RemoveAt(0);
 	}
-
-	
 }
